refactor: move startup helpers
